+++ packages/rainbowkit/src/components/ConnectButton/ConnectButtonRenderer.tsx
@@ -87,14 +87,27 @@ export function ConnectButtonRenderer({
     showRecentTransactions;
 
   const { showBalance } = useShowBalance();
-  const shouldShowBalance = showBalance
-    ? normalizeResponsiveValue(showBalance)[
+
+  const computeShouldShowBalance = () => {
+    if (typeof showBalance === 'boolean') {
+      return showBalance;
+    }
+
+    if (showBalance) {
+      return normalizeResponsiveValue(showBalance)[
         isMobile() ? 'smallScreen' : 'largeScreen'
-      ]
-    : true;
+      ];
+    }
+
+    return true;
+  };
+
+  const shouldShowBalance = computeShouldShowBalance();
+
   const { data: balanceData } = useRealtimeBalance(
     shouldShowBalance ? address : undefined,
   );
+
   const displayBalance = balanceData
     ? `${abbreviateETHBalance(parseFloat(balanceData.formatted))} ${
         balanceData.symbol