+++ packages/rainbowkit/src/components/ConnectOptions/DesktopOptions.tsx
@@ -76,7 +76,12 @@ export function DesktopOptions({ onClose }: { onClose: () => void }) {
 
   const { connector } = useContext(WalletButtonContext);
 
-  const mergeEIP6963WithRkConnectors = true;
+  // The `WalletButton` component made the connect modal appear empty when trying to connect.
+  // This happened because of a mix up between EIP-6963 and RainbowKit connectors.
+  // The problem was finding the correct `wallet.id`. `WalletButton` uses RainbowKit's id,
+  // but EIP-6963 uses `rdns` for its id. We now don't merge EIP-6963 and RainbowKit
+  // connectors if user interacts with `WalletButton` component.
+  const mergeEIP6963WithRkConnectors = !connector;
 
   const wallets = useWalletConnectors(mergeEIP6963WithRkConnectors)
     .filter((wallet) => wallet.ready || !!wallet.extensionDownloadUrl)