+++ packages/rainbowkit/src/components/RainbowKitProvider/ModalContext.tsx
@@ -31,12 +31,16 @@ interface ModalContextValue {
   openAccountModal?: () => void;
   openChainModal?: () => void;
   openConnectModal?: () => void;
+  isWalletConnectModalOpen: boolean;
+  setIsWalletConnectModalOpen: (isWalletConnectModalOpen: boolean) => void;
 }
 
 const ModalContext = createContext<ModalContextValue>({
   accountModalOpen: false,
   chainModalOpen: false,
   connectModalOpen: false,
+  isWalletConnectModalOpen: false,
+  setIsWalletConnectModalOpen: () => {},
 });
 
 interface ModalProviderProps {
@@ -61,6 +65,9 @@ export function ModalProvider({ children }: ModalProviderProps) {
     isModalOpen: chainModalOpen,
     openModal: openChainModal,
   } = useModalStateValue();
+
+  const [isWalletConnectModalOpen, setIsWalletConnectModalOpen] =
+    useState(false);
 
   const connectionStatus = useConnectionStatus();
 
@@ -104,6 +111,7 @@ export function ModalProvider({ children }: ModalProviderProps) {
           accountModalOpen,
           chainModalOpen,
           connectModalOpen,
+          isWalletConnectModalOpen,
           openAccountModal:
             isCurrentChainSupported && connectionStatus === 'connected'
               ? openAccountModal
@@ -115,6 +123,7 @@ export function ModalProvider({ children }: ModalProviderProps) {
             connectionStatus === 'unauthenticated'
               ? openConnectModal
               : undefined,
+          setIsWalletConnectModalOpen,
         }),
         [
           connectionStatus,
@@ -125,6 +134,7 @@ export function ModalProvider({ children }: ModalProviderProps) {
           openChainModal,
           openConnectModal,
           isCurrentChainSupported,
+          isWalletConnectModalOpen,
         ],
       )}
     >
@@ -157,7 +167,19 @@ export function useChainModal() {
   return { chainModalOpen, openChainModal };
 }
 
+export function useWalletConnectOpenState() {
+  const { isWalletConnectModalOpen, setIsWalletConnectModalOpen } =
+    useContext(ModalContext);
+
+  return { isWalletConnectModalOpen, setIsWalletConnectModalOpen };
+}
+
 export function useConnectModal() {
   const { connectModalOpen, openConnectModal } = useContext(ModalContext);
-  return { connectModalOpen, openConnectModal };
+  const { isWalletConnectModalOpen } = useWalletConnectOpenState();
+
+  return {
+    connectModalOpen: connectModalOpen || isWalletConnectModalOpen,
+    openConnectModal,
+  };
 }