+++ packages/rainbowkit/src/wallets/walletConnectors/index.ts
@@ -2,6 +2,7 @@ import { argentWallet } from './argentWallet/argentWallet';
 import { bifrostWallet } from './bifrostWallet/bifrostWallet';
 import { bitgetWallet } from './bitgetWallet/bitgetWallet';
 import { bitskiWallet } from './bitskiWallet/bitskiWallet';
+import { bloomWallet } from './bloomWallet/bloomWallet';
 import { braveWallet } from './braveWallet/braveWallet';
 import { clvWallet } from './clvWallet/clvWallet';
 import { coin98Wallet } from './coin98Wallet/coin98Wallet';
@@ -15,6 +16,7 @@ import { frameWallet } from './frameWallet/frameWallet';
 import { frontierWallet } from './frontierWallet/frontierWallet';
 import { imTokenWallet } from './imTokenWallet/imTokenWallet';
 import { injectedWallet } from './injectedWallet/injectedWallet';
+import { kresusWallet } from './kresusWallet/kresusWallet';
 import { ledgerWallet } from './ledgerWallet/ledgerWallet';
 import { metaMaskWallet } from './metaMaskWallet/metaMaskWallet';
 import { mewWallet } from './mewWallet/mewWallet';
@@ -25,6 +27,8 @@ import { oneKeyWallet } from './oneKeyWallet/oneKeyWallet';
 import { phantomWallet } from './phantomWallet/phantomWallet';
 import { rabbyWallet } from './rabbyWallet/rabbyWallet';
 import { rainbowWallet } from './rainbowWallet/rainbowWallet';
+import { ramperWallet } from './ramperWallet/ramperWallet';
+import { roninWallet } from './roninWallet/roninWallet';
 import { safeWallet } from './safeWallet/safeWallet';
 import { safeheronWallet } from './safeheronWallet/safeheronWallet';
 import { safepalWallet } from './safepalWallet/safepalWallet';
@@ -45,6 +49,7 @@ export {
   bifrostWallet,
   bitgetWallet,
   bitskiWallet,
+  bloomWallet,
   braveWallet,
   clvWallet,
   coin98Wallet,
@@ -58,6 +63,7 @@ export {
   frontierWallet,
   imTokenWallet,
   injectedWallet,
+  kresusWallet,
   ledgerWallet,
   metaMaskWallet,
   mewWallet,
@@ -68,6 +74,8 @@ export {
   phantomWallet,
   rabbyWallet,
   rainbowWallet,
+  ramperWallet,
+  roninWallet,
   safeWallet,
   safeheronWallet,
   safepalWallet,